+++ src/main/java/woojooin/planit/domain/goal/isa/mapper/IsaAccountMapper.java
@@ -1,13 +1,13 @@
-package woojooin.planit.domain.object.isa.mapper;
+package woojooin.planit.domain.goal.isa.mapper;
 
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
-import woojooin.planit.domain.object.isa.dto.req.IsaAccountProductEditReq;
-import woojooin.planit.domain.object.isa.dto.req.IsaAccountProductRegisterReq;
-import woojooin.planit.domain.object.isa.dto.res.IsaAccountProductRes;
+import woojooin.planit.domain.goal.isa.dto.req.IsaAccountProductEditReq;
+import woojooin.planit.domain.goal.isa.dto.req.IsaAccountProductRegisterReq;
+import woojooin.planit.domain.goal.isa.dto.res.IsaAccountProductRes;
 
 @Mapper
 public interface IsaAccountMapper {
@@ -23,4 +23,6 @@ public interface IsaAccountMapper {
 
 	void softDelete(@Param("memberId") Long memberId, @Param("requestList") List<IsaAccountProductEditReq> editReqs);
 	void upsert(@Param("memberId") Long memberId, @Param("requestList") List<IsaAccountProductEditReq> editReqs);
+
+	Long getTotalIsaProfitByMemberId(@Param("memberId") Long memberId);
 }
