+++ src/main/java/woojooin/planit/global/response/ResponseCode.java
@@ -46,6 +46,7 @@ public enum ResponseCode {
 	ISA_REGISTRATION_FAILED("ISA-010", "상품 등록에 실패했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
 	ISA_UPDATE_FAILED("ISA-011", "상품 수정에 실패했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
 	ISA_DELETE_FAILED("ISA-012", "상품 삭제에 실패했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
+	ISA_TAX_CALCULATION_FAILED("ISA-013", "세금 절약 금액 계산에 실패했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
 
 
 	// Deposit Account Domain Errors
